+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Reconcile individual clusterRoleBinding for each org namespace to grant automation access to silences.
+- Reconcile individual clusterRoleBinding for each org namespace to grant and revoke automation access to silences.
 
 ## [0.29.0] - 2022-05-30